change data type of max

+++ 0x14-bit_manipulation/2-get_bit.c
@@ -4,11 +4,12 @@
  * get_bit - function returns the value of a bit at a given index
  * @n: integer value
  * @index: index to look for starting from 0
+ * Return: value of the bit or -1 if error
 */
 
 int get_bit(unsigned long int n, unsigned int index)
 {
-	unsigned int max = 0x01;
+	unsigned long int max = 0x01;
 
 	max << index;
 	if(max == 0)
